[C++ API] Move impl of CudaMappingOptions to .cc
This changeset moves the implementation of CUDA backend specific options
to .cc. This is a general pattern in the refactoring where backend-specific
implementations are fully specialized in .cc files in the proper backend and
templated types and functions live in tc/core and are implemented in -inl.h
files.
